Generalizing time series simulation time check from `PrimaryServiceWorker`
